+++ lib/finders/astar-finder.ts
@@ -75,9 +75,7 @@ export class AStarFinder {
       !this.grid.isWalkableAt(endPosition) ||
       !this.grid.isWalkableAt(startPosition)
     ) {
-      console.log(
-        'Path could not be created because the start and/or end position is/are not walkable.'
-      );
+      // Path could not be created because the start and/or end position is/are not walkable.
       return [];
     }
 
@@ -116,7 +114,7 @@ export class AStarFinder {
     // As long the open list is not empty, continue searching a path
     while (this.openList.length !== 0) {
       // Get node with lowest f value
-      let currentNode = _.minBy(this.openList, function(o) {
+      let currentNode = _.minBy(this.openList, (o) => {
         return o.getFValue();
       });
 
@@ -174,7 +172,7 @@ export class AStarFinder {
         }
       }
     }
-    console.log('Path could not be created. ');
+    // Path could not be created
     return [];
   }